requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
art==6.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
autopep8==2.0.2
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
boto3==1.26.16
botocore==1.29.16
build==1.2.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
config==0.5.1
contourpy==1.2.1
coverage==7.2.7
cryptography==41.0.1
cursor==1.3.5
cycler==0.10.0
Cython==3.0.10
debugpy==1.8.13
decorator==5.1.1
decouple==0.0.7
deep-sort-realtime==1.3.2
defusedxml==0.7.1
diffusers==0.23.0
dnspython==2.3.0
docopt==0.6.2
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.115.11
fastapi-cli==0.0.7
FastAPI-SQLAlchemy==0.2.1
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.6.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fpdf==1.7.2
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.19.0
idna==2.10
imageio==2.33.0
importlib-metadata==5.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.1.2
jax==0.4.29
jaxlib==0.4.29
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
jsonify==0.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.15.0
kiwisolver==1.4.5
lapx==0.5.9.post1
lazy_loader==0.3
libclang==16.0.6
lxml==5.2.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
multi-key-dict==2.0.3
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
opencv-contrib-python==4.10.0.82
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
openpyxl==3.1.5
opentype-feature-freezer==1.32.2
opt-einsum==3.3.0
orjson==3.10.15
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pbr==5.11.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pushbullet.py==0.9.1
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pybind11==2.12.0
pybind11_global==2.12.0
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pygame==2.5.2
Pygments==2.17.2
pymongo==4.3.3
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyOpenSSL==23.2.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
python-calamine==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jenkins==1.8.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2024.1
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.35.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
roboflow==1.1.32
rpds-py==0.18.0
rsa==4.9
s3transfer==0.6.0
safetensors==0.4.0
sahi==0.11.16
samila==1.1
scikit-image==0.22.0
scipy==1.13.1
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.5
SQLAlchemy==2.0.15
stack-data==0.6.3
starlette==0.46.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threaded==4.2.0
tifffile==2023.9.26
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.35
ultralytics-thop==2.0.0
urllib3==1.26.16
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==15.0.1
Werkzeug==2.3.4
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.2.2
yarg==0.1.9
yolov5==7.0.13
yolov7==0.0.1
zipp==3.9.0
